Update of folder structure.

+++ go.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [[ -z "$JENKINS_HOME" ]]; then
+    echo "JENKINS_HOME is not defined" 1>&2
+    exit 1
+fi
+
 mvn install || { echo "Build failed"; exit 1; }
 
 rm -rf $JENKINS_HOME/plugins/analysis-core*
